requirements.txt
fastapi
uvicorn
python-dotenv
psycopg2-binary
peewee
pydantic[email]
passlib[bcrypt]
python-jose[cryptography]
python-multipart
pytest
requests